+++ CHANGELOG.md
@@ -26,6 +26,8 @@ _The format is based on [Keep a Changelog](http://keepachangelog.com/) and this 
    
   `keyframes` no longer returns an animation name, instead it returns an object which has method `.getName()` for the purpose of getting the animation name.
 
+- Migrate to use new `React.forwardRef` API, by [@probablyup](https://github.com/probablyup); note that this removes the `innerRef` API since it is no longer needed.
+
 ## [v3.4.2] - 2018-08-07
 
 - Fix a regression from [#1843](https://github.com/styled-components/styled-components/pull/1892) that breaks deferred injection and duplicates rules, by [@kitten](https://github.com/kitten) (see [#1892](https://github.com/styled-components/styled-components/pull/1892))